+++ tidy3d/components/geometry/utils_2d.py
@@ -12,9 +12,10 @@ from ..scene import Scene
 from ..structure import Structure
 
 
-def increment_float(val: np.float32, sign) -> np.float32:
-    """Applies a small positive or negative shift to a 32bit float using numpy.nextafter,"""
-    """but additionally handles some corner cases."""
+def increment_float(val: float, sign) -> float:
+    """Applies a small positive or negative shift as though `val` is a 32bit float
+    using numpy.nextafter, but additionally handles some corner cases.
+    """
     # Infinity is left unchanged
     if val == inf or val == -inf:
         return val
@@ -23,6 +24,11 @@ def increment_float(val: np.float32, sign) -> np.float32:
         sign = 1
     else:
         sign = -1
+
+    # Avoid small increments within subnormal values
+    if np.abs(val) <= np.finfo(np.float32).tiny:
+        return val + sign * np.finfo(np.float32).tiny
+
     # Numpy seems to skip over the increment from -0.0 and +0.0
     # which is different from c++
     val_inc = np.nextafter(val, sign * inf, dtype=np.float32)